Split api utils
